Update object loading (#32)

+++ src/render/opengl/opengl_renderer_api.hpp
@@ -24,7 +24,7 @@ class OpenGLRendererAPI : public RendererAPI
    Clear() override;
 
    void
-   DrawIndexed(const std::shared_ptr< VertexArray >& vertexArray, uint32_t indexCount = 0) override;
+   DrawIndexed(const std::shared_ptr< VertexArray >& vertexArray, size_t indexCount = 0) override;
 
    void
    DrawLines(uint32_t count) override;
